# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API de Autenticación de Usuarios",
      "version": "1.0.0",
      "description": "Esta API permite registrar usuarios, iniciar sesión y acceder a zonas protegidas mediante autenticación JWT.",
      "contact": {
        "email": "tuemail@dominio.com"
      }
    },
    "paths": {
      "/api/register": {
        "post": {
          "tags": ["Auth"],
          "summary": "Registrar nuevo usuario",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "nombre",
                    "apellido",
                    "fecha_nacimiento",
                    "genero",
                    "correo",
                    "telefono",
                    "contrasena",
                    "rol_id"
                  ],
                  "properties": {
                    "nombre": {"type": "string"},
                    "apellido": {"type": "string"},
                    "fecha_nacimiento": {"type": "string", "format": "date"},
                    "genero": {"type": "string"},
                    "correo": {"type": "string", "format": "email"},
                    "telefono": {"type": "string"},
                    "contrasena": {"type": "string"},
                    "rol_id": {"type": "integer"}
                  }
                }
              }
            }
          },
          "responses": {
            "201": {"description": "Usuario registrado correctamente"},
            "400": {"description": "Error de validación"},
            "500": {"description": "Error del servidor"}
          }
        }
      },
      "/login": {
        "post": {
          "tags": ["Autenticación"],
          "summary": "Inicio de sesión de usuarios",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["correo", "contrasena"],
                  "properties": {
                    "correo": {"type": "string", "format": "email"},
                    "contrasena": {"type": "string"}
                  }
                }
              }
            }
          },
          "responses": {
            "200": {"description": "Inicio de sesión exitoso"},
            "401": {"description": "Credenciales inválidas"}
          }
        }
      },
      "/api/protected": {
        "get": {
          "tags": ["Auth"],
          "summary": "Acceso protegido mediante token JWT",
          "parameters": [
            {
              "name": "Authorization",
              "in": "header",
              "description": "Token JWT en formato Bearer",
              "required": true,
              "schema": {"type": "string"}
            }
          ],
          "responses": {
            "200": {"description": "Acceso concedido a zona protegida"},
            "401": {"description": "Token inválido o no encontrado"}
          }
        }
      }
    }
  }